requirements.txt
playwright==1.41.2
python-dotenv==1.0.0 